snapshot prune: check that the elements match instead of equals
The property map in the results list is not necessarily ordered.
Instead of making sure that we have an equal object for the prune,
let's just check if the elements match, since that's what we care
about.  This fixes some flakyness because of map ordering.

Note that we don't need to do the same for snap.Resources, and it's
better that way, because we care about the ordering of the resources
in the snapshot.