[sdk/python] Send plain values to the engine in output invokes and keep track of input dependencies/secrets (#17460)
### Description 

Resolves #17459
And the Python portion of #17474

When python output invokes serialized their inputs, if any of these
inputs is an output, the shape of the output is maintained in the
serialized form across the wire. However, for invokes we shouldn't
maintain the shape of the output values, instead the values sent across
the wire should be plain. Flipping the option `keep_output_values` to
`False` fixes the problem.

Also here we keep track of input dependencies for output invokes and
merge the full list of dependencies from all inputs into the
dependencies list of the result of the invoke.