[pcl] Detect and error on binding component nodes that self-reference their source (#17538)
### Description

Recently we have been encountering out of memory errors when trying to
convert terraform modules. Found out that the problem is two-fold:
- pulumi-terraform-converter rewriting `module example "../../"` (tf)
into `component example "./."` (pcl)
- we try to bind component `example` but it reads its source from the
same directory as the program of the component, causing an infinite
recursion.

This PR fixes part 2 of the issue, erroring out if the component's
source is the same as the current binder directory path.

Partially addresses
https://github.com/pulumi/pulumi-converter-terraform/issues/194 we still
need to fix part 1 on the converter side of things