Add a conformance test for preserving map keys (#17350)
This adds a conformance test checking we can use map keys like "MY_KEY"
and maintain their casing through program gen. For Python we need to
simplify the type passed to genObjectConsExpression so that we can
reliably determine the target type and check if it's a `TypedDict` or
not, otherwise we might treat it as a plain object and preserve the
keys, instead of `PyName`ing them.

This deletes the "typeddict" test from the codegen tests. Firstly this
is now covered by these conformance tests, secondly it wasn't actually
valid because the schema it was testing against wasn't a valid schema,
and the program snapshot was only how it was because of a bug in
programgen.

That bug is now fixed in this change which fixes
https://github.com/pulumi/pulumi/issues/17294.

---------

Co-authored-by: Julien Poissonnier <julien@caffeine.lu>