Port over developer documentation on Pulumi types (#17544)
Pulumi is built on top of a core type system that is loosely "specified"
by the gRPC/Protocol buffer definitions we expose and a set of core
implementations in e.g. the Pulumi engine. Historically we documented
this type system as part of the old developer documentation. This commit
ports that content to the new documentation, tidying it up and
correcting things that have since changed/may no longer be the case.