refactor InputPropertyErrors to be part of the rpcerror package
We already had a package that was appropriate for these errors, and it
makes sense to extend that to also include the InputPropertiesErrors,
instead of having them live somewhere else.  This is technically a
breaking change for component provider authors, but
InputPropertiesErrors have only been introduced recently, and have not
been widely advertised, so I think it's preferrable to have this
cleanup now rather than having to live with the two packages that do
similar things forever.