token_source: show logs to users in case of an error
We're trying to figure out why this token sometimes expires and fails
to renew in https://github.com/pulumi/pulumi/issues/7094.  However
doing so is proving tricky, as the issue is hard to reproduce.
Ideally we'd get some logs from users that did manage to reproduce the
issue.

Since the issue only happens occasionally it's hard to reproduce
though, and most users don't run pulumi with logging enabled.  Let's
make collecting and submitting the logs easier for users by collecting
the logs internally, and always printing them in case an error
happens, or if the token is expired.