Add Uv as a Python toolchain
We currently support Pip and Poetry as toolchains for Python. This PR adds Uv as a third option.

The Uv toolchain manages virtualenvs, and we do not allow using Uv with a global Python installation. If the `virtualenv` runtime option is not set, we default to creating a virtualenv named `.venv` next to the nearest `uv.lock` or `pyproject.toml`.