Support generating random resources in lifecycle tests
Snapshot integrity errors are very problematic when they occur and can
be hard to spot and prevent. To this end, #17213 outlines a plan to
introduce [fuzzing](https://en.wikipedia.org/wiki/Fuzzing) to our suite
of lifecycle tests in order to find cases and executions which might
violate snapshot integrity. This commit kicks this off by introducing a
`fuzzing` package to the suite and adding types and generators (from the
`pgregory.net/rapid` library) for generating random resources. The idea
is that from resources we can progress to generating random snapshots;
from there to programs and provider configurations and so on; and with
all these pieces execute random tests in an attempt to find snapshot
integrity bugs before our customers do.

Part of #17213