Support generating random providers in lifecycle tests
Snapshot integrity errors are very problematic when they occur and can
be hard to spot and prevent. To this end, #17213 outlines a plan to
introduce [fuzzing](https://en.wikipedia.org/wiki/Fuzzing) to our suite
of lifecycle tests in order to find cases and executions which might
violate snapshot integrity. This commit extends the `fuzzing` package of
the suite to support generating random providers. A provider may
configure operations such as `Create`, `Diff`, etc. to fail or succeed
in a number of ways (e.g. no diff, changes, replaces, etc.) on a per-URN
basis, so as to exercise a variety of code paths in step
generation/execution.

Part of #17213