Support generating random snapshots in lifecycle tests
Snapshot integrity errors are very problematic when they occur and can
be hard to spot and prevent. To this end, #17213 outlines a plan to
introduce [fuzzing](https://en.wikipedia.org/wiki/Fuzzing) to our suite
of lifecycle tests in order to find cases and executions which might
violate snapshot integrity. This commit extends the `fuzzing` package of
the suite to support generating random valid snapshots, the idea being
that this will enable us to generate good starting states for our
property tests. A snapshot contains a number of random resources which
may depend on one another in valid ways. Snapshots may also contain
deleted resources which have not yet been cleaned from the snapshot.

Part of #17213