Support generating random programs in lifecycle tests
Snapshot integrity errors are very problematic when they occur and can
be hard to spot and prevent. To this end, #17213 outlines a plan to
introduce [fuzzing](https://en.wikipedia.org/wiki/Fuzzing) to our suite
of lifecycle tests in order to find cases and executions which might
violate snapshot integrity. This commit extends the `fuzzing` package of
the suite to support generating random programs. A program is based upon
a snapshot and may randomly append and prepend new resources and copy,
drop or update existing ones.

Part of #17213