Add the ability to fuzz parents and aliases
In #17623 and the PRs that followed it, we added fuzz testing
capabilities to the engine's lifecycle test suite, with a view to
randomly generating test cases in the hopes of proactively finding
snapshot integrity bugs in our code. This commit extends the fuzzing
library to generate a reproducing lifecycle test case in the event that
a snapshot integrity error is found, hopefully aiding in debugging and
pinning down the exact cause of the error. Code is written to a file in
a temporary directory, which may optionally be overridden using an
environment variable. This might be useful in e.g. a GitHub action that
fuzzes periodically so that failing cases' reproductions can be made
available as artifacts for download.

Part of #17213