Support generating random fixtures in lifecycle tests (#17627)
Snapshot integrity errors are very problematic when they occur and can
be hard to spot and prevent. To this end, #17213 outlines a plan to
introduce [fuzzing](https://en.wikipedia.org/wiki/Fuzzing) to our suite
of lifecycle tests in order to find cases and executions which might
violate snapshot integrity. This commit extends the `fuzzing` package of
the suite to support generating random fixtures and adds documentation
for the tactics employed when doing so.

A fixture comprises an initial snapshot, a program to run against that
initial snapshot, a set of providers to use during that program's
execution, and an operation to run. Fixtures can then be generated as
part of a Rapid property test, allowing us to fuzz random combinations
in order to hunt down snapshot integrity issues.

Part of #17213