Add the ability to fuzz parents and aliases (#17654)
In #17623 and the PRs that followed it, we added fuzz testing
capabilities to the engine's lifecycle test suite, with a view to
randomly generating test cases in the hopes of proactively finding
snapshot integrity bugs in our code. This commit extends the fuzzer so
that it randomly parents, re-parents, and aliases reparented resources,
covering the various parent/child relationships that these actions lead
to and which can result in snapshot integrity issues if handled
improperly. In particular, we can now fuzz the following:

* Randomly parenting a resource to another resource in an initial
snapshot.
* Randomly updating a resource in a program to either add, remove, or
change an existing parent.
* Randomly aliasing a resource whose parent (and consequently URN) has
been changed by a program to point back to the URN it had originally.

Part of #17213