build release binaries in parallel with other CI steps
In CI we want to have binaries that emit coverage and have race
detection built in.  However we don't want that in release builds, as
neither of these things is useful there, and is just slowing down the
binary.  We currently achieve this by rebuilding the binaries after
all the tests have run.

However building the binaries takes additional time (~3min currently).
Wit the introduction of the artifact-suffix option in
`ci-build-binaries` we can now do better, and build the release
binaries in parallel with running the tests.  Do that here to save
some time when running CI.