Fix language conformance tests
In #17589 we added the `package` block to PCL in order to support
parameterized providers. As part of this work, we added an initial set
of language conformance tests for parameterized providers. Presently
there are some workarounds involved in doing so as historically we have
only tested non-parameterized providers. This commit introduces one that
we missed, without which some assertions currently fail since they are
unaware of parameterized providers.