Update pulumi --help documentation to use bucketv2/BucketV2 (#17692)
As part of https://github.com/pulumi/home/issues/3583 documentation is
updated to recommend BucketV2 usage over Bucket. Part of the
documentation is generated automatically from `pulumi --help`. To get
this right the Pulumi CLI itself needs to be updated. This change
updates the example used in `pulumi --help` to use BucketV2 instead of
Bucket.