remove prepare-release step from on-pr workflow (#17674)
Currently we have a prepare-release step in the on-pr workflow. This
step updates the draft release that we're using to release
pulumi/pulumi. It is only run when the `ci/test` label is added to a PR,
which is done very rarely, but definitely has happened before.

This could in theory lead to the wrong/unreviewed code being released,
if someone adds a ci/test label to a PR, and no other PR is merged
between that and a freeze PR is merged (note that the freeze PR will not
update the draft release, but will create the next release).

I think our prepare-release workflow is stable enough that we don't need
to test it in CI at this point, and we should never run the risk of
releasing unreviewed code. Therefore remove this step completely from
the on-pr CI workflow.