simplify the keep running in the service backend
Simplify keeping the machine running in the service backend.  Instead
of splitting it up, we can simply make sure to keep the computer
running for the whole apply.  This also removes the risk of the
resetSleep function being nil, and potentially panic'ing.  (This is
currently not possible, but the setup we currently have is somewhat
brittle)