Bump gRPC dependencies and migrate `grpc.Dial` (#17701)
We'd like to bump `pulumi-java` to 0.17.0, but in order to do so we'll
need to first bump our gRPC dependencies to match. Doing so incurs
dealing with a deprecation -- `grpc.Dial` has been deprecated in favour
of `grpc.NewClient`. This commit makes that upgrade and migration
separately so that we can bump Java in a subsequent piece of work.