stop the token source loop correctly on 403 errors
In the token source loop, when a 403 error happens, there's no way
it's going to succeed in the future, and we want to abort it.  We try
to do so currently, however we're passing the incorrect type to
`errors.As`, and thus never get an expiredTokenError, and never abort
the loop.

Fix that, and add a test that reproduces the whole round trip through
the client by setting up a fake http server.