stop the token source loop correctly on 403 errors (#17714)
In the token source loop, when a 403 error happens, there's no way it's
going to succeed in the future, and we want to abort it. We try to do so
currently, however we're passing the incorrect type to `errors.As`, and
thus never get an expiredTokenError, and never abort the loop.

Note that we are no longer retrying this in a tight loop, but if someone
calls `GetToken` frequently for whatever reason, we'd still end up in
the same situation as in https://github.com/pulumi/home/issues/3660.

Fix that, and add a test that reproduces the whole round trip through
the client by setting up a fake http server.

Fixes https://github.com/pulumi/pulumi/issues/17551